Fixes after the review
